Moved `emit_event` to module event_utils.py
tendrl-bug-id: Tendrl/specifications#167
